fix del animals when del aviary/ do aviary.class in separate file
